@@ -1,11 +1,11 @@
 name: animated_search_bar
 description: A Beautiful and Simple SearchBar widget with animation transition. It can be fully customized with label, labelStyle, searchDecoration, etc. It also maintains onChange state.
-version: 1.3.0
+version: 2.0.0
 author: Mudassir<hey.mudassir@gmail.com>
 homepage: https://github.com/ukieTux/animated_search_bar
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.20.0 <2.0.0"
 
 dependencies:
